ArchID  💜🧡❤️💛💚💙
🌐 https://archid.app/ 🌐

ArchID is who you are in the Archway Network 🌟. A name for your address and so much more.

A social security number is how official sources identify you 🆔, but not how you introduce yourself. You do that with your name. Start using your ArchID instead of an unmemorable string to identify yourself in Archway.